Adjust lnc(4) for 64-bit platforms should it get newbus'ified.

+++ sys/dev/lnc/if_lncvar.h
@@ -183,7 +183,7 @@ struct nic_info {
 	int ident;         /* Type of card */
 	int ic;            /* Type of ic, Am7990, Am79C960 etc. */
 	int mem_mode;
-	int iobase;
+	long iobase;
 	int mode;          /* Mode setting at initialization */
 };
